+++ httpdns-20160201/ChangeLog.md
@@ -1,3 +1,9 @@
+2024-08-29 Version: 2.0.1
+- Update API ListDomains: add param Search.
+- Update API ListDomains: update param PageSize.
+- Update API ListDomains: update response param.
+
+
 2021-12-06 Version: 2.0.0
 - First build sdk in AMP.